+++ future/src/ct/ct_codebox.h
@@ -24,35 +24,10 @@
 #include <gtkmm.h>
 #include <gtksourceviewmm.h>
 #include "ct_const.h"
-#include "ct_main_win.h"
+#include "ct_widgets.h"
+
 
 enum class CtPixTabCBox : int {Pixbuf=0, Table=1, CodeBox=2};
-
-class CtAnchoredWidget : public Gtk::EventBox
-{
-public:
-    CtAnchoredWidget(const int charOffset, const std::string& justification);
-    virtual ~CtAnchoredWidget();
-
-    void insertInTextBuffer(Glib::RefPtr<Gsv::Buffer> rTextBuffer);
-    Glib::RefPtr<Gtk::TextChildAnchor> getTextChildAnchor() { return _rTextChildAnchor; }
-
-    virtual void applyWidthHeight(const int parentTextWidth) = 0;
-    virtual void to_xml(xmlpp::Element* /*p_node_parent*/, const int offset_adjustment) { _charOffset += offset_adjustment; }
-
-    void updateOffset(int charOffset) { _charOffset = charOffset; }
-    void updateJustification(std::string justification) { _justification = justification; }
-
-    int getOffset() { return _charOffset; }
-    const std::string& getJustification() { return _justification; }
-
-protected:
-    Gtk::Frame _frame;
-    Gtk::Label _labelWidget;
-    int _charOffset;
-    std::string _justification;
-    Glib::RefPtr<Gtk::TextChildAnchor> _rTextChildAnchor;
-};
 
 class CtTextCell
 {
